+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [v0.2] - 2020-04-15
 
 ### Added
-- Best support for sub queries filters : when using the WHERE or HAVING clause, the prefix PATABASE_COLUMN_LITERALL in the filter value allows to match with the result of the main query (insead of a non dynamic value).
+- Best support for sub queries filters : when using the WHERE or HAVING clause, the prefix _PATABASE_COLUMN_LITERALL_ in the filter value allows to match with the result of the main query (insead of a non dynamic value). 
 
 ### Changed
 - **Possible break change** Drop support for php < 7.1